+++ src/features/better-layers/main.cpp
@@ -6,7 +6,8 @@
 #include "include.hpp"
 
 using namespace geode::prelude;
-using namespace nwo5::syntax;
+using namespace nwo5::ui::prelude;
+using namespace nwo5::utils::setup;
 
 class $modify(BetterLayersEditorUI, EditorUI) {
     struct Fields {
@@ -20,6 +21,7 @@ class $modify(BetterLayersEditorUI, EditorUI) {
 
         // remind me to make some kind of system of select filters in api cuz this is copy pasted from replace obj
         bool canDestroyUndo = false;
+        bool updateLockButton = true;
     };
 
     static constexpr float GAP = 5.0f;
@@ -44,133 +46,119 @@ class $modify(BetterLayersEditorUI, EditorUI) {
 
         fields->settings = std::make_unique<BetterLayers::LayerSettings>();
 
-        fields->layerInput = nwo5::utils::createTextInput(LAYER_INPUT_SIZE.width, LAYER_INPUT_SIZE.height, "All", [this] (const std::string& pStr) {
-            if (pStr == "a" || pStr == "A") {
-                editor::setLayer(editor::constants::ALL_LAYERS);
-            }
-            else if (pStr.contains('A') || pStr.contains('l')) {
-                m_fields->layerInput->setString("");
-            }
-            else if (!pStr.empty()) {
-                auto res = utils::numFromString<int>(pStr);
+        fields->layerInput = ui::node(
+            Setup(ui::input(LAYER_INPUT_SIZE, "All"))
+                .id("layer-input"_spr)
+                .filter("aA1234567890")
+                .maxCharCount(4)
+                .callback([this] (const std::string& pStr) {
+                    if (pStr == "a" || pStr == "A") {
+                        editor::setLayer(editor::constants::ALL_LAYERS);
+                    }
+                    else if (pStr.contains('A') || pStr.contains('l')) {
+                        m_fields->layerInput->setString(string::remove(pStr, "Al"));
+                    }
+                    else if (!pStr.empty()) {
+                        auto res = utils::numFromString<int>(pStr);
 
-                if (auto res = utils::numFromString<int>(pStr); res.isOk()) {
-                    editor::setLayer(std::clamp(res.unwrap(), 0, editor::constants::MAX_LAYERS));
-                }
-            }
-        });
-        
-        fields->layerInput->setFilter("aA1234567890");
-        fields->layerInput->setMaxCharCount(4);
-
-        auto nextLayer = nwo5::utils::setupNode(
-            nwo5::utils::createButtonFrame("GJ_arrow_03_001.png", this, menu_selector(EditorUI::onGroupUp)),
-
-            SetNodeID{"next-layer-button"_spr},
-            SetNodeScaleWithSize{LAYER_SHIFT_BUTTON_SIZE}
+                        if (auto res = utils::numFromString<int>(pStr); res.isOk()) {
+                            editor::setLayer(std::clamp(res.unwrap(), 0, editor::constants::MAX_LAYERS));
+                        }
+                    }
+                })
         );
 
-        nextLayer->setRotationY(180.0f);
-
-        auto prevLayer = nwo5::utils::setupNode(
-            nwo5::utils::createButtonFrame("GJ_arrow_03_001.png", this, menu_selector(EditorUI::onGroupDown)),
-
-            SetNodeID{"prev-layer-button"_spr},
-            SetNodeScaleWithSize{LAYER_SHIFT_BUTTON_SIZE}
+        auto nextLayer = ui::node(
+            Setup(ui::buttonFrame("GJ_arrow_03_001.png", this, menu_selector(EditorUI::onGroupUp)))
+                .id("next-layer-button"_spr)
+                .scaleToFit(LAYER_SHIFT_BUTTON_SIZE)
+                .flipX()
+        );
+        auto prevLayer = ui::node(
+            Setup(ui::buttonFrame("GJ_arrow_03_001.png", this, menu_selector(EditorUI::onGroupDown)))
+                .id("prev-layer-button"_spr)
+                .scaleToFit(LAYER_SHIFT_BUTTON_SIZE)
         );
 
-        fields->allLayersButton = nwo5::utils::setupNode( // girl robtop, ur function names, what the fuck is this, why only name it layer here </3
-            nwo5::utils::createButtonFrame("GJ_arrow_02_001.png", this, menu_selector(EditorUI::onGoToBaseLayer)),
-
-            SetNodeID{"next-free-layer-button"_spr},
-            SetNodeScaleWithSize{LAYER_EXTRA_BUTTON_SIZE}
+        fields->allLayersButton = ui::node(
+            // girl robtop, ur function names, what the fuck is this, why only name it layer here </3
+            Setup(ui::buttonFrame("GJ_arrow_02_001.png", this, menu_selector(EditorUI::onGoToBaseLayer)))
+                .id("next-free-layer-button"_spr)
+                .scaleToFit(LAYER_EXTRA_BUTTON_SIZE)
         );
 
-        auto nextFreeLayer = nwo5::utils::setupNode(
-            nwo5::utils::createButtonFrame("GJ_plusBtn_001.png", this, menu_selector(BetterLayersEditorUI::onNextFreeLayer)),
-
-            SetNodeID{"next-free-layer-button"_spr},
-            SetNodeScaleWithSize{LAYER_EXTRA_BUTTON_SIZE},
-            SetNodeVisibility{Settings::BetterLayers::nextFreeButton.get()}
+        auto nextFreeLayer = ui::node(
+            Setup(ui::buttonFrame("GJ_plusBtn_001.png", this, menu_selector(BetterLayersEditorUI::onNextFreeLayer)))
+                .id("next-free-layer-button"_spr)
+                .scaleToFit(LAYER_EXTRA_BUTTON_SIZE)
+                .visible(Settings::BetterLayers::nextFreeButton.get())
         );
 
-        fields->lockLayerButton = nwo5::utils::setupNode(
-            CCMenuItemToggler::create(
-                CCSprite::createWithSpriteFrameName("warpLockOffBtn_001.png"),
-                CCSprite::createWithSpriteFrameName("warpLockOnBtn_001.png"),
-                this, menu_selector(BetterLayersEditorUI::onToggleLayerLocked)
-            ),
-
-            SetNodeID{"lock-layer-button"_spr},
-            SetNodeScaleWithSize{LAYER_EXTRA_BUTTON_SIZE},
-            SetNodeVisibility{Settings::BetterLayers::lockButton.get()}
+        fields->lockLayerButton = ui::node(
+            Setup(ui::togglerFrame("warpLockOffBtn_001.png", "warpLockOnBtn_001.png", this, menu_selector(BetterLayersEditorUI::onToggleLayerLocked)))
+                .id("lock-layer-button"_spr)
+                .scaleToFit(LAYER_EXTRA_BUTTON_SIZE)
+                .visible(Settings::BetterLayers::lockButton.get())
         );
 
-        auto layerSettings = nwo5::utils::setupNode(
-            nwo5::utils::createButtonFrame("GJ_optionsBtn_001.png", this, menu_selector(BetterLayersEditorUI::onLayerSettings)),
-
-            SetNodeID{"layer-settings-button"_spr},
-            SetNodeScaleWithSize{(LAYER_SHIFT_BUTTON_SIZE + LAYER_EXTRA_BUTTON_SIZE) / 2}
+        auto layerSettings = ui::node(
+            Setup(ui::buttonFrame("GJ_optionsBtn_001.png", this, menu_selector(BetterLayersEditorUI::onLayerSettings)))
+                .id("layer-settings-button"_spr)
+                .scaleToFit((LAYER_SHIFT_BUTTON_SIZE + LAYER_EXTRA_BUTTON_SIZE) / 2)
         );
 
         auto oldLayerMenu = m_currentLayerLabel->getParent();
-
         oldLayerMenu->setVisible(false);
+
         if (m_currentLayerLabel) {
-            m_currentLayerLabel->setScale(0.0f); // be compat
+            m_currentLayerLabel->setPositionX(999.0f); // i give up dealing with better edits editablelabelproxy fuck that
         }
 
-        fields->newLayerMenu = nwo5::utils::setupNode(
-            CCMenu::create(),
-
-            SetNodeID{"new-layer-menu"_spr},
-            SetNodeScale{oldLayerMenu},
-            SetNodeAnchor{oldLayerMenu},
-            SetNodePosition{oldLayerMenu},
-            SetNodeOrder{oldLayerMenu->getZOrder()},
-            SetNodeChildren{
-                nextFreeLayer,
-                nextLayer,
-                fields->layerInput,
-                prevLayer,
-                fields->allLayersButton,
-                fields->lockLayerButton,
-                layerSettings
-            },
-            SetNodeParent{this}
+        fields->newLayerMenu = ui::node(
+            Setup(ui::menu(ui::horizontalDistrbLayout(GAP)))
+                .id("new-layer-menu"_spr)
+                .scale(oldLayerMenu)
+                .anchor(oldLayerMenu)
+                .pos(oldLayerMenu)
+                .order(oldLayerMenu)
+                .children(
+                    layerSettings,
+                    fields->lockLayerButton,
+                    fields->allLayersButton,
+                    prevLayer,
+                    fields->layerInput,
+                    nextLayer,
+                    nextFreeLayer
+                )
+                .parent(this)
         );
 
         m_uiItems->addObject(fields->newLayerMenu);
 
-        fields->newLayerMenu->setLayout(RowLayout::create()
-            ->setGap(GAP)
-            ->setGrowCrossAxis(false)
-            ->setAxisAlignment(AxisAlignment::Center)
-            ->setAxisReverse(true)
-            ->setAutoGrowAxis(true)
-            ->setAutoScale(false)
-        );
-
-        updateGroupIDLabel();
+        m_editorLayer->m_currentLayer = m_editorLayer->m_level->m_lastBuildGroupID;
+        updateLayerMenu();
 
         return true;
     }
 
     void updateLayerMenu() {
-        auto fields = m_fields.self();;
+        auto fields = m_fields.self();
 
         if (!fields->newLayerMenu) {
             return;
         }
 
-        if (const auto layer = editor::currentLayer(); layer == editor::constants::ALL_LAYERS) {
+        // editor technically not initialized yet when im calling in init so i cant call thru my api (mayb ill fix this eventually)
+        const auto layer = m_editorLayer->m_currentLayer;
+
+        if (layer == editor::constants::ALL_LAYERS) {
             fields->layerInput->setString("All");
         }
         else {
             fields->layerInput->setString(nwo5::utils::numToString(layer));
         }
         
-        if (editor::layerLocked(editor::currentLayer())) {
+        if (editor::layerLocked(layer)) {
             fields->layerInput->getInputNode()->getTextLabel()->setColor(ccYELLOW);
         }
         else {
@@ -180,11 +168,13 @@ class $modify(BetterLayersEditorUI, EditorUI) {
         fields->lockLayerButton->setVisible(
             Settings::BetterLayers::lockButton.get() 
             && GameManager::get()->getGameVariable(GameVar::LayerLocking) 
-            && editor::currentLayer() != editor::constants::ALL_LAYERS
+            && layer != editor::constants::ALL_LAYERS
         );
-        fields->lockLayerButton->toggle(editor::layerLocked(editor::currentLayer()));
+        if (fields->updateLockButton) {
+            fields->lockLayerButton->toggle(editor::layerLocked(layer));
+        }
         
-        fields->allLayersButton->setVisible(editor::currentLayer() != editor::constants::ALL_LAYERS);
+        fields->allLayersButton->setVisible(layer != editor::constants::ALL_LAYERS);
 
         fields->newLayerMenu->updateLayout();
     }
@@ -218,7 +208,11 @@ class $modify(BetterLayersEditorUI, EditorUI) {
     }
 
     void onToggleLayerLocked(CCObject*) {
+        m_fields->updateLockButton = false;
+
         editor::lockLayer(editor::currentLayer(), !editor::layerLocked(editor::currentLayer()));
+        
+        m_fields->updateLockButton = true;
     }
 
     void updateGroupIDLabel() {