+++ src/features/shared.cpp
@@ -3,7 +3,7 @@
 #include "features/shared.hpp"
 
 using namespace geode::prelude;
-using namespace nwo5::syntax;
+using namespace nwo5::utils::setup;
 
 class $modify(SharedLevelEditorLayer, LevelEditorLayer) {
     struct Fields {